backend/requirements.txt
arrow==1.2.2
asgiref==3.5.0
blessed==1.19.1
Django==4.0.4
django-filter==21.1
django-picklefield==3.0.1
django-q==1.3.9
django-spa==0.3.6
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
gunicorn==20.1.0
lxml==4.8.0
PyJWT==2.3.0
pyphen==0.12.0
python-dateutil==2.8.2
python-docx==0.8.11
pytz==2022.1
redis==3.5.3
six==1.16.0
sqlparse==0.4.2
wcwidth==0.2.5
whitenoise==5.3.0
